Moving function to TreeMapChartData and setting getColor function
